refactor for the context handler 1 of 2
