+++ src/Server.js
@@ -34,7 +34,8 @@ import SynapsDB from './services/synapsdb/index.js';
 // Managers
 //import SessionManager from './managers/session/index.js';
 import LayerManager from './managers/layer/index.js';
-import TreeManager from './managers/tree/index.js';
+import TreeManager from './managers/contextTree/index.js';
+import WorkspaceManager from './managers/workspace/index.js';
 
 import setupTransportsConfig from './transports/setupTransportConfig.js';
 
@@ -90,12 +91,14 @@ const DEFAULT_TRANSPORTS = {
 };
 
 // Managers
+const workspaceManager = new WorkspaceManager(indexManager.createIndex('workspaces'));
 const layerManager = new LayerManager(indexManager.createIndex('layers'));
-
 const treeManager = new TreeManager({
     treeIndex: indexManager.createIndex('tree'),
     layerIndex: layerManager
 });
+
+const contextTree = treeManager.createContextTree();
 
 /**
  * Canvas Server
@@ -452,6 +455,8 @@ export {
     logger,
     indexManager,
     db,
+    contextTree,
+    workspaceManager,
 };
 
 export default Server;
