+++ src/managers/contextTree/lib/Tree.js
