+++ src/managers/contextTree/lib/TreeNode.js
