+++ src/managers/workspace/lib/Workspace.js
